+++ framework/CHANGELOG.md
@@ -54,6 +54,7 @@ Yii Framework 2 Change Log
 - Enh #16126: Allows to configure `Connection::dsn` by config array (leandrogehlen)
 - Chg #11397: `yii\i18n\MessageFormatter` polyfills and `yii\i18n\MessageFormatter::parse()` method were removed resulting in performance boost. See UPGRADE for compatibility notes (samdark)
 - Chg #16247: Cloning components will now clone their behaviors as well (brandonkelly)
+- Enh #16487: Added circular reference detection in DI container (hiqsol)
 
 2.0.14.2 under development
 ------------------------